requirements.txt
Django==1.10.3
whitenoise==2.0.6
Pillow==3.4.2
https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow-0.11.0-cp35-cp35m-linux_x86_64.whl